Use :latest trait to explicity set latest_edition_id
Previously every `Edition` created by FactoryBot was using an
`after_create` callback to set its associated document's
`latest_edition_id` to itself.

This does not reflect any such automatic behaviour early in
the edition's lifecycle in the codebase. Currently
`Document#latest_edition_id` is set in two places at the end
of the edition's "workflow" lifecycle:

1. by the [`ScheduleEditionService`][], which calls
  `Edition#update_document_reference_to_latest_edition!`

2. by the [`PublishEditionService`][] which manually sets
  `latest_edition_id`

To help show that this `Document#latest_edition` association
is set manually we move the `Edition` factory's `after_create`
callback to a [trait][] so its use is visible and explicit.

In the next commit we'll add a clearer description of
`Document#latest_edition`.

[`ScheduleEditionService`]:
https://github.com/alphagov/content-block-manager/blob/c898be18ca81f89a3e1d6aacfc32c609714828c9/app/services/schedule_edition_service.rb#L8-L15

[`PublishEditionService`]:
https://github.com/alphagov/content-block-manager/blob/8eff034e7c1a50277ac31892db36cec9751f9ee2/app/services/publish_edition_service.rb#L51-L56

[trait]:
https://thoughtbot.github.io/factory_bot/traits/summary.html