Allow preview to be navigated via Turbo
Currently, when a user clicks on a link or submits a form, this triggers
a complete page load, which means the user gets sent to the top of the
page each time and has to scroll down. This hooks into Turbo to listen
to link clicks or form submits and load the iframe content via Turbo.
We also alter the cursor so the user gets feedback that the next page
is loading.

Because we’re using an iframe this doesn’t just work off the bat, and
we have to listen to `click` or `submit` events on the iframe itself.